+++ CHANGES.md
@@ -22,6 +22,9 @@ Version Next
 
 * The clientId has been removed from the string representation of internal broker objects.
 
+* Afkak no longer logs a warning if a response is received for a request that was previously cancelled.
+  When a truly unexpected response is received (e.g., due to a broker bug) an error is logged with up to 1024 bytes of the response.
+
 Version 3.0.0
 =============